IB/core: Change sm_lid from 16 to 32 bits.
Change sm_lid from 16 to 32 bits in the kernel. The UABI is unchanged
and still returns 16 bit sm_lids. SA is aware of the full 32 bit
sm_lid by using the IB_AH_GRH flags for its address handle and stoing
the sm_lid in the GID entry.


Conflicts:
	include/rdma/ib_addr.h

+++ drivers/staging/rdma/hfi1/verbs.h
@@ -412,7 +412,7 @@ void hfi1_rc_hdrerr(
 
 u8 ah_to_sc(struct ib_device *ibdev, struct ib_ah_attr *ah_attr);
 
-struct ib_ah *hfi1_create_qp0_ah(struct hfi1_ibport *ibp, u16 dlid);
+struct ib_ah *hfi1_create_qp0_ah(struct hfi1_ibport *ibp, u32 dlid);
 
 void hfi1_rc_rnr_retry(unsigned long arg);
 void hfi1_add_rnr_timer(struct rvt_qp *qp, u32 to);
